hooks added to make code more manageable
