+++ web/package.json
@@ -99,8 +99,10 @@
     "dotenv": "^16.4.5",
     "fontawesome": "^5.6.3",
     "graphql": "^16.9.0",
+    "html2canvas": "^1.4.1",
     "immer": "^10.0.4",
     "jest-environment-jsdom": "^29.7.0",
+    "jspdf": "^2.5.2",
     "minimist": "1.2.8",
     "multiparty": "^4.2.3",
     "node-fetch": "^3.3.2",
